Refactored singular and empy files
