+++ learnrl/agents/__init__.py
@@ -1,5 +1,5 @@
 # LearnRL a python library to learn and use reinforcement learning
 # Copyright (C) 2020 Mathïs FEDERICO <https://www.gnu.org/licenses/>
 
-from learnrl.agents.standard.agent import StandardAgent
-from learnrl.agents.random.agent import RandomAgent
+from learnrl.agents.standard import StandardAgent
+from learnrl.agents.random import RandomAgent
