--- learnrl/environments/multi_agents/nim/__init__.py
@@ -1,2 +1,0 @@
-# LearnRL a python library to learn and use reinforcement learning
-# Copyright (C) 2020 Mathïs FEDERICO <https://www.gnu.org/licenses/>
