--- learnrl/environments/multi_agents/nim/game.py
@@ -1,8 +1,0 @@
-# LearnRL a python library to learn and use reinforcement learning
-# Copyright (C) 2020 Mathïs FEDERICO <https://www.gnu.org/licenses/>
-
-from gym import Env
-import gym.spaces as spaces
-
-class NimEnv(Env):
-    pass
